etl/requirements.txt
pymupdf>=1.24.0
pytesseract>=0.3.10
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
jsonschema>=4.17.0
opencv-python-headless>=4.8.0
